move utility functions and classes to core subpackage
