+++ simphony_mayavi/core/api.py
@@ -5,9 +5,14 @@ from .doc_utils import mergedocs
 from .constants import CELL2VTKCELL, FACE2VTKCELL, EDGE2VTKCELL
 from .constants import VTKCELLTYPES, VTKFACETYPES, VTKEDGETYPES
 from .constants import ELEMENT2VTKCELLTYPES, VTKCELLTYPE2ELEMENT
+from .cell_array_tools import cell_array_slicer
+from .cuba_data_accumulator import CUBADataAccumulator, gather_cells
+from .cuba_data_extractor import CUBADataExtractor
 
 __all__ = [
     "CubaData", "supported_cuba", "CellCollection", "mergedocs",
     "CELL2VTKCELL", "FACE2VTKCELL", "EDGE2VTKCELL",
     "VTKCELLTYPES", "VTKFACETYPES", "VTKEDGETYPES",
-    "ELEMENT2VTKCELLTYPES", "VTKCELLTYPE2ELEMENT"]
+    "ELEMENT2VTKCELLTYPES", "VTKCELLTYPE2ELEMENT",
+    "cell_array_slicer", "CUBADataAccumulator", "CUBADataExtractor",
+    "gather_cells"]
