+++ simphony_mayavi/core/cell_array_tools.py
