+++ simphony_mayavi/sources/api.py
@@ -1,14 +1,8 @@
 from .particles_source import ParticlesSource
 from .lattice_source import LatticeSource
 from .mesh_source import MeshSource
-from .cuds_data_accumulator import CUDSDataAccumulator
-from .cuds_data_extractor import CUDSDataExtractor
-from .utils import cell_array_slicer
 
 __all__ = [
     'ParticlesSource',
     'LatticeSource',
-    'MeshSource',
-    'CUDSDataExtractor',
-    'CUDSDataAccumulator',
-    'cell_array_slicer']
+    'MeshSource']
